% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_frac.R
\name{frac_by_top}
\alias{frac_by_top}
\title{Create proportions, by top_party, of split/straight/etc}
\usage{
frac_by_top(var, tbl = vdf)
}
\arguments{
\item{var}{Name of office to subset to, e.g. "USH"}

\item{tbl}{Vote object, in long form. Can be arrow connection}
}
\description{
Create proportions, by top_party, of split/straight/etc
}
