% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/viz_waffle.R
\name{gg_wfl_long}
\alias{gg_wfl_long}
\title{Waffle Plot of votes, taking long data}
\usage{
gg_wfl_long(
  tbl,
  var,
  nrows = 31,
  rev = FALSE,
  office_nam = NULL,
  blank = FALSE,
  legend = FALSE,
  title_size = 0.8,
  check_ncand = TRUE
)
}
\arguments{
\item{var}{The variable in the dataset to use as the colors}

\item{nrows}{number of rows and columns}

\item{office_nam}{Title of the office. If NULL, use recode_abbrv of variable name}
}
\description{
Waffle Plot of votes, taking long data
}
